requirements.txt
fastapi
uvicorn
pytest
PyMuPDF
requests
numpy
faiss-cpu